[Guardian] Lift verified S3 read layer from monitor into hashi-guardian
Move `GuardianPollerCore` + `GuardianLogDir` (the reusable, attestation-
and signature-verified S3 log reader) out of `hashi-monitor`'s
`rpc/guardian.rs` into a new `hashi-guardian::s3_reader` module, next to
the `S3Logger` it wraps. This is the read counterpart to the existing
write layer, and it's the shared core the upcoming KP/operator init
tooling will consume — neither can depend on monitor-private code.

The monitor-domain wrapper (`GuardianWithdrawalsPoller`,
`VerifiedWithdrawal`, and the `VerifiedLogRecord -> MonitorEvent`
mapping) stays in `hashi-monitor` over the relocated core. The init
consumers (`heartbeat_checks`, `limiter_recovery`) now import the core
from `hashi-guardian`.

The core's only tie to monitor was `now_unix_seconds`; it now uses the
equivalent `hashi_types::guardian::time_utils::now_timestamp_secs`, so
the lift carries zero monitor coupling. Pure relocation, no behavior
change.
