+++ crates/hashi-monitor/src/provisioner/heartbeat_checks.rs
@@ -2,9 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::domain::now_unix_seconds;
-use crate::rpc::guardian::GuardianLogDir;
-use crate::rpc::guardian::GuardianPollerCore;
 use hashi_guardian::s3_logger::S3Logger;
+use hashi_guardian::s3_reader::GuardianLogDir;
+use hashi_guardian::s3_reader::GuardianPollerCore;
 use hashi_types::guardian::LogMessage;
 use hashi_types::guardian::VerifiedLogRecord;
 use hashi_types::guardian::time_utils::UnixSeconds;
